requirements.txt
pyyaml>=6.0.1
fpdf2>=2.8.0
